requirement.txt
django==
djangorestframework==
pillow==
mysql==